Rename all bin and BIN to mig and MIG, at least in zargon.cpp. Remember the idea is that mig (migrate) is either a binary offset or a native pointer. For the moment it's still only a binary offset, but hopefully the next commit will be different. BTW we checked and defining mig_t as either uint16_t or uint32_t worked fine (with warnings, but somewhat better perf (47K+ not 46K). But have reverted to uint64_t due to that having the same size as native pointers.

+++ src/bridge.cpp
@@ -243,7 +243,7 @@ std::string mem_dump()
             case xM1:        nlines=2;   break;
             case xMLIST:     nlines=2;   break;
         }
-        int len = v.first-offset;
+        int len = (int)(v.first-offset);
         int blank_lines = 0;
         line.clear();
         for( int lines=0, i=0; v.first!=xBOARDA && i<len; i++ )
@@ -318,7 +318,7 @@ std::string mem_dump()
                 line.clear();
             }
         }
-        offset = v.first;
+        offset = (int)v.first;
         if( need_newline)
         {
             if( blank_lines > 0 )
